LPS-112102 move preview API
