+++ modules/apps/headless/headless-admin-content/headless-admin-content-api/bnd.bnd
@@ -1,6 +1,4 @@
 Bundle-Name: Liferay Headless Admin Content API
 Bundle-SymbolicName: com.liferay.headless.admin.content.api
 Bundle-Version: 1.0.0
-Export-Package:\
-	com.liferay.headless.admin.content.dto.v1_0,\
-	com.liferay.headless.admin.content.resource.v1_0
+Export-Package: com.liferay.headless.admin.content.resource.v1_0
